Split ContextPlug into 3 separate plugs
