+++ test/sanbase_web/graphql/blockchain_address/blockchain_address_transactions_api_test.exs
@@ -117,7 +117,7 @@ defmodule SanbaseWeb.Graphql.BlockchainAddressTransactionsApiTest do
         },
         "toAddress" => %{
           "address" => "0xf4b51b14b9ee30dc37ec970b50a486f37686e2a8",
-          "labels" => [%{"name" => "whale"}, %{"name" => "centralized_exchange"}]
+          "labels" => [%{"name" => "centralized_exchange"}, %{"name" => "whale"}]
         },
         "trxHash" => "0x4bd5d44da7f69c5227530728249431072087b5f9780eec704869fc768922121e",
         "trxValue" => 888_888.0
@@ -134,7 +134,7 @@ defmodule SanbaseWeb.Graphql.BlockchainAddressTransactionsApiTest do
         },
         "toAddress" => %{
           "address" => "0xf4b51b14b9ee30dc37ec970b50a486f37686e2a8",
-          "labels" => [%{"name" => "whale"}, %{"name" => "centralized_exchange"}]
+          "labels" => [%{"name" => "centralized_exchange"}, %{"name" => "whale"}]
         },
         "trxHash" => "0x21a56440bedb1a5c2f4adca4a6f9fbccf13bd741d63c0e3b2214a6ee418a5974",
         "trxValue" => 5.5,
